Get rules configuration in JSON
